Refactor: Extract wstring functions into a separate header
No extra includes as it's currently inlined anyway (which is probably a good idea to keep unless there's a significant reason to change that).
